+++ src/main/java/com/squareup/connect/models/OrderLineItemDiscount.java
@@ -125,10 +125,10 @@ public class OrderLineItemDiscount {
   }
 
    /**
-   * The discount's Unique identifier, unique only within this order. This field is read-only.
+   * Unique ID that identifies the discount only within this order.  This field is read-only.
    * @return uid
   **/
-  @ApiModelProperty(value = "The discount's Unique identifier, unique only within this order. This field is read-only.")
+  @ApiModelProperty(value = "Unique ID that identifies the discount only within this order.  This field is read-only.")
   public String getUid() {
     return uid;
   }