+++ src/main/java/com/squareup/connect/models/OrderReturnLineItemModifier.java
@@ -51,10 +51,10 @@ public class OrderReturnLineItemModifier {
   }
 
    /**
-   * The return modifier's Unique identifier, unique only within this order
+   * Unique ID that identifies the return modifier only within this order.  This field is read-only.
    * @return uid
   **/
-  @ApiModelProperty(value = "The return modifier's Unique identifier, unique only within this order")
+  @ApiModelProperty(value = "Unique ID that identifies the return modifier only within this order.  This field is read-only.")
   public String getUid() {
     return uid;
   }