+++ src/main/java/com/squareup/connect/models/OrderRoundingAdjustment.java
@@ -42,10 +42,10 @@ public class OrderRoundingAdjustment {
   }
 
    /**
-   * The rounding adjustment's Unique identifier, unique only within this order. This field is read-only.
+   * Unique ID that identifies the rounding adjustment only within this order.  This field is read-only.
    * @return uid
   **/
-  @ApiModelProperty(value = "The rounding adjustment's Unique identifier, unique only within this order. This field is read-only.")
+  @ApiModelProperty(value = "Unique ID that identifies the rounding adjustment only within this order.  This field is read-only.")
   public String getUid() {
     return uid;
   }